ARCH-3: Created StubStructures file and moved StubRadioButtonItem there
